requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
chromadb==0.5.5
pypdf==4.3.1
sentence-transformers==3.1.1
pydantic-settings==2.5.2
PyJWT==2.9.0
